+++ src/openfe/protocols/openmm_afe/abfe_units.py
@@ -26,7 +26,7 @@ from openmmtools.states import ThermodynamicState
 from rdkit import Chem
 
 from openfe.protocols.openmm_afe.equil_afe_settings import (
-    ABFEBoreschRestraintSettings,
+    BoreschRestraintSettings,
     SettingsBaseModel,
 )
 from openfe.protocols.openmm_utils import system_validation
@@ -218,7 +218,7 @@ class ABFEComplexSetupUnit(ComplexComponentsMixin, ComplexSettingsMixin, BaseAbs
         guest_atom_ids: list[int],
         host_atom_ids: list[int],
         temperature: Quantity,
-        settings: ABFEBoreschRestraintSettings,
+        settings: BoreschRestraintSettings,
     ) -> tuple[BoreschRestraintGeometry, BoreschRestraint]:
         """
         Get a Boresch-like restraint Geometry and OpenMM restraint force
@@ -236,7 +236,7 @@ class ABFEComplexSetupUnit(ComplexComponentsMixin, ComplexSettingsMixin, BaseAbs
           A list of atom indices defining the host molecules in the universe.
         temperature : openff.units.Quantity
           The temperature of the simulation where the restraint will be added.
-        settings : ABFEBoreschRestraintSettings
+        settings : BoreschRestraintSettings
           Settings on how the Boresch-like restraint should be defined.
 
         Returns
@@ -249,17 +249,23 @@ class ABFEComplexSetupUnit(ComplexComponentsMixin, ComplexSettingsMixin, BaseAbs
         # Take the minimum of the two possible force constants to check against
         frc_const = min(settings.K_thetaA, settings.K_thetaB)
 
+        if settings.guest_restraint_ids is not None:
+            guest_restraint_atoms_idxs = list(settings.guest_restraint_ids)
+        else:
+            guest_restraint_atoms_idxs = None
+
+        if settings.host_restraint_ids is not None:
+            host_restraint_atoms_idxs = list(settings.host_restraint_ids)
+        else:
+            host_restraint_atoms_idxs = None
+
         geom = geometry.boresch.find_boresch_restraint(
             universe=universe,
             guest_rdmol=guest_rdmol,
             guest_idxs=guest_atom_ids,
             host_idxs=host_atom_ids,
-            guest_restraint_atoms_idxs=list(settings.guest_restraint_ids)
-            if settings.guest_restraint_ids is not None
-            else None,
-            host_restraint_atoms_idxs=list(settings.host_restraint_ids)
-            if settings.host_restraint_ids is not None
-            else None,
+            guest_restraint_atoms_idxs=guest_restraint_atoms_idxs,
+            host_restraint_atoms_idxs=host_restraint_atoms_idxs,
             host_selection=settings.host_selection,
             anchor_finding_strategy=settings.anchor_finding_strategy,
             dssp_filter=settings.dssp_filter,
