+++ src/openfe/protocols/openmm_afe/equil_afe_settings.py
@@ -217,82 +217,6 @@ class ABFEPreEquilOutputSettings(MDOutputSettings):
         return v
 
 
-class ABFEBoreschRestraintSettings(BoreschRestraintSettings):
-    """
-    Settings to define a Boresch-style restraint between
-    two groups of atoms named ``host`` and ``guest``.
-
-    The restraint is defined in the following manner::
-
-      H2                         G2
-       ⧹                         ⧸
-        ⧹                       ⧸
-         H1 ——— H0 ——— G0 ——— G1
-
-    Where HX represents the X index of ``host_atoms``
-    and GX the X index of ``guest_atoms``.
-
-    By default, the Boresch-like restraint will be
-    obtained using a modified version of the
-    search algorithm implemented by Baumann et al. [1].
-
-    If ``guest_atoms`` and ``host_atoms`` are defined,
-    these indices will be used instead.
-
-    References
-    ----------
-    [1] Baumann, Hannah M., et al. "Broadening the scope of binding free
-        energy calculations using a Separated Topologies approach." (2023).
-    [2] Wu, Zhiyi, et al. "Optimizing Absolute Binding Free Energy
-        Calculations for Production Usage."
-        (2025; DOI 10.26434/chemrxiv-2025-q08ld-v2)
-    [3] Alibay, Irfan, et al. "Evaluating the use of absolute binding
-        free energy in the fragment optimisation process" (2022).
-    """
-
-    host_restraint_ids: tuple[int, int, int] | None = None
-    """
-    The indices of the host component atoms to restrain.
-    The entries define the H0, H1, and H2 atoms in order.
-    If defined, these will override any automatic selection.
-
-    Notes
-    -----
-    These indices refer to the atom index in the host component
-    (e.g. the PDB file defining the host) and not the final
-    index of the fully solvated complex which will be simulated.
-    """
-    guest_restraint_ids: tuple[int, int, int] | None = None
-    """
-    The indices of the guest component atoms to restraint.
-    The entries define the G0, G1, and G2 atoms in order.
-    If defined, these will override any automatic selection.
-
-    Notes
-    -----
-    These indices refer to the atom index in the SmallMoleculeComponent
-    defining the guest (e.g. in the input SDF file) and not the final
-    index of the fully sollvated complex which will be simulated.
-    """
-
-    @field_validator("guest_restraint_ids", "host_restraint_ids")
-    def positive_idxs_three_tuple(cls, v):
-        if v is not None:
-            if any([i < 0 for i in v]):
-                errmsg = "``guest_atoms`` and ``host_atoms`` cannot have negative indices."
-                raise ValueError(errmsg)
-        return v
-
-    @model_validator(mode="after")
-    def check_restraint_ids_defined(self):
-        if (self.host_restraint_ids is None) ^ (self.guest_restraint_ids is None):
-            errmsg = (
-                "`guest_restraint_ids` and `host_restraint_ids` must both "
-                "either be defined or undefined"
-            )
-            raise ValueError(errmsg)
-
-
 # This subclasses from SettingsBaseModel as it has vacuum_forcefield and
 # solvent_forcefield fields, not just a single forcefield_settings field
 class AbsoluteSolvationSettings(SettingsBaseModel):
