+++ src/openfe/tests/protocols/openmm_abfe/test_abfe_settings.py
@@ -7,7 +7,6 @@ from openfe.protocols.openmm_afe import (
     AbsoluteBindingProtocol,
     AbsoluteBindingSettings,
 )
-from openfe.protocols.openmm_afe.equil_afe_settings import ABFEBoreschRestraintSettings
 
 
 @pytest.fixture()
@@ -69,31 +68,6 @@ def test_monotonic_lambda_windows(val, default_settings):
         lambda_settings.lambda_elec = val["elec"]
         lambda_settings.lambda_vdw = val["vdw"]
         lambda_settings.lambda_restraints = val["restraints"]
-
-
-@pytest.mark.parametrize("parameter", ["host_restraint_ids", "guest_restraint_ids"])
-def test_boresch_restraints_restraint_negative_ids(parameter):
-    setting = ABFEBoreschRestraintSettings()
-
-    errmsg = "``guest_atoms`` and ``host_atoms`` cannot have negative indices."
-    with pytest.raises(ValueError, match=errmsg):
-        setattr(setting, parameter, [1, 2, -3])
-
-
-@pytest.mark.parametrize("parameter", ["host_restraint_ids", "guest_restraint_ids"])
-def test_boresch_restraints_too_many_ids(parameter):
-    setting = ABFEBoreschRestraintSettings()
-
-    errmsg = "Tuple should have at most 3 items after validation, not 4"
-    with pytest.raises(ValueError, match=errmsg):
-        setattr(setting, parameter, [1, 2, 3, 4])
-
-
-def test_boresch_restraint_partially_defined_ids():
-    setting = ABFEBoreschRestraintSettings()
-
-    with pytest.raises(ValueError, match="must both either be defined or undefined"):
-        setting.host_restraint_ids = [1, 2, 3]
 
 
 def test_equil_not_all_complex(default_settings):
